make vgem_test for drm-tests
In order to test the VGEM kernel module, a userspace program was made to
test it. This is that program. It was used in development of the kernel
module but is being uploaded in the hopes of someday being useful in the
Chrome OS tests.

TEST=vgem_test
BUG=chromium:426172

Change-Id: Ic11c723500d61cfdd82a1e2cb00191b8d8d030d5
Reviewed-on: https://chromium-review.googlesource.com/230907
Reviewed-by: Ilja Friedel <ihf@chromium.org>
Commit-Queue: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>